Yongyong-Agentic-Core/requirements.txt
fastapi
uvicorn
sqlalchemy
opencv-python
numpy
easyocr
rembg
python-dotenv
google-generativeai
tavily-python
llama-index-core>=0.10.0
llama-index-graph-stores-neo4j>=0.7.0
llama-index-llms-gemini>=0.1.0
llama-index-vector-stores-chroma>=0.2.0
neo4j
groq
chromadb